distribute_to_pi/requirements.txt
audioread==2.1.8
Keras
tensorflow
Pillow
sounddevice